Passing tests with @michwill's scoring suggestions.
